Fixes:    starting icc support Task:     114862 Details:  I would not want to support it yet, there is some uic3 uglyness to work on 	  but in so far as running the code through another compiler it did catch a 	  few legitimate errors.
[git-p4: depot-paths = "//depot/qt/main/": change = 220442]

+++ src/gui/graphicsview/qgraphicsitem.cpp
@@ -1709,7 +1709,7 @@ void QGraphicsItem::setData(int key, const QVariant &value)
 */
 int QGraphicsItem::type() const
 {
-    return UserType;
+    return (int)UserType;
 }
 
 /*!
